refactor(search): split 875-LOC search_content into widget part files
Extracts search bar, sort, filter panel, filter chips, result card, and
empty/header state widgets into lib/views/search/widgets/ part files per
File_Organization_Rules R1. Each file now under the 300 LOC hard limit.
